# Refactor HivePartitionFunction in preparation for supporting complex types